refactor: user에서 auth로 패키지명 변경
