+++ src/main/java/roomescape/domain/auth/service/UserService.java
@@ -1,0 +1,5 @@
+package roomescape.domain.auth.service;
+
+public class UserService {
+
+}
