--- src/main/java/roomescape/domain/user/controller/UserController.java
@@ -1,5 +1,0 @@
-package roomescape.domain.user.controller;
-
-public class UserController {
-
-}
