--- src/main/java/roomescape/domain/user/entity/User.java
@@ -1,5 +1,0 @@
-package roomescape.domain.user.entity;
-
-public class User {
-
-}
