--- src/main/java/roomescape/domain/user/service/UserService.java
@@ -1,5 +1,0 @@
-package roomescape.domain.user.service;
-
-public class UserService {
-
-}
