+++ src/main/java/roomescape/ui/ViewController.java
@@ -35,4 +35,9 @@ public class ViewController {
     public String reservation() {
         return "reservation";
     }
+
+    @GetMapping("/login")
+    public String login() {
+        return "login";
+    }
 }
