+++ src/test/java/roomescape/domain/reservation/entity/ReservationTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import roomescape.common.exception.InvalidArgumentException;
-import roomescape.domain.user.entity.Name;
+import roomescape.domain.auth.entity.Name;
 
 class ReservationTest {
